Bug Fixes v1.0
Fixed cards that can only target self. They should now be able to be played and apply their affects accordingly. Also hopefully fixed the end turn bug that was sometimes encountered in combat. Believe it was because some enemy action do not "target" and it was leading to null exceptions. Finally, fixed a few card scriptable objects to make sure they have the proper data.